Remove default arguments from signals
This is done because it didn't work in the first place and it seems not supported anymore anyways. I found no way to make it work and the cpp bindings for example don't seem to support it either. The same can be achieved anyways through binds in the connect method.